refactor: add item settings with maxWidth
