+++ src/services/file/interfaces/extra.ts
@@ -1,5 +1,12 @@
 import { ItemType } from '../../../constants';
 import { UnknownExtra } from '../../../interfaces';
+import { ItemSettings } from '@/services/items/interfaces/itemSettings';
+
+export enum MaxWidth {
+  Small = 'xs',
+  Medium = 'md',
+  Large = 'lg',
+}
 
 /**
  * @deprecated Use FileItemProperties instead
@@ -27,3 +34,7 @@ export interface S3FileItemExtra extends UnknownExtra {
 }
 
 export type FileItemExtra = S3FileItemExtra | LocalFileItemExtra;
+
+export interface FileItemSettings extends ItemSettings {
+  maxWidth?: MaxWidth | false;
+}
