+++ src/services/items/interfaces/item.ts
@@ -7,31 +7,14 @@ import {
 } from '../../../interfaces/extra';
 import { AppItemExtra } from '../../app/';
 import { EtherpadItemExtra } from '../../etherpad';
-import { LocalFileItemExtra, S3FileItemExtra } from '../../file';
+import {
+  FileItemSettings,
+  LocalFileItemExtra,
+  S3FileItemExtra,
+} from '../../file';
 import { H5PItemExtra } from '../../h5p';
-import { CCLicenseAdaptions, Member, OldCCLicenseAdaptations } from '@/index';
-
-export interface ItemSettings {
-  lang?: string;
-  isPinned?: boolean;
-  showChatbox?: boolean;
-  hasThumbnail?: boolean;
-  isResizable?: boolean;
-  isCollapsible?: boolean;
-  enableSaveActions?: boolean;
-  tags?: string[];
-  displayCoEditors?: boolean;
-  ccLicenseAdaption?:
-    | `${CCLicenseAdaptions}`
-    | CCLicenseAdaptions
-    // todo: these are the old licenses, we might remove them at some point.
-    | `${OldCCLicenseAdaptations}`;
-}
-
-export interface EmbeddedLinkItemSettings extends ItemSettings {
-  showLinkIframe?: boolean;
-  showLinkButton?: boolean;
-}
+import { EmbeddedLinkItemSettings, ItemSettings } from './itemSettings';
+import { Member } from '@/index';
 
 export interface Item<S = ItemSettings> {
   id: string;
@@ -65,11 +48,11 @@ export type EmbeddedLinkItemType<S = ItemSettings> = {
   extra: EmbeddedLinkItemExtra;
   settings: EmbeddedLinkItemSettings;
 } & Item<S>;
-export type LocalFileItemType<S = ItemSettings> = {
+export type LocalFileItemType<S = FileItemSettings> = {
   type: `${ItemType.LOCAL_FILE}`;
   extra: LocalFileItemExtra;
 } & Item<S>;
-export type S3FileItemType<S = ItemSettings> = {
+export type S3FileItemType<S = FileItemSettings> = {
   type: `${ItemType.S3_FILE}`;
   extra: S3FileItemExtra;
 } & Item<S>;
